fix: code refactoring for stats
